feat: 리프레시토큰 테스트 중에서 데이터 수정중

+++ src/main/java/com/example/todolist_backend/dto/todo/ToDoCreateRequest.java
@@ -15,7 +15,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor // 파라미터 없는 기본 생성자 자동으로 생성
 @Getter // 필드의 getter 메서드 자동 생성
 public class ToDoCreateRequest { // 사용자 요청 -> db 에 이 내용으로 저장
-     private Long userId;
+    private int userId;
+//    private Long userId;
     private String content;
     private boolean done;
