+++ src/main/java/com/example/todolist_backend/repository/UserRepository.java
@@ -7,7 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
+//public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUserName(String userName);
     // Optional<User> findByAccount(String account);
     boolean existsByAccountAndPassword(String account, String password); // 인터페이스는 기본적으로 public 접근제어자를 가지므로, 명시적으로 public 사용 불필요
